imp: Improve the initial content and layout
