+++ .pre-commit-config.yaml
@@ -19,7 +19,7 @@ repos:
     rev: "v3.0.2"
     hooks:
       - id: prettier
-        exclude: "build|theme|pages/dashboards"
+        exclude: "build|theme|pages/projects"
 
   - repo: local
     hooks:
