+++ mkdocs.yml
@@ -23,7 +23,7 @@ extra_javascript:
 
 nav:
   - Home: "index.md"
-  - Dashboards: "dashboards/index.md"
+  - Projects: "projects/index.md"
   - Mentors: "mentors/index.md"
   - Donate: "donate/index.md"
