--- pages/dashboards/README.md
@@ -1,3 +1,0 @@
-The content of this folder is created automatically.
-
-Don't change it manually.
