--- pages/dashboards/example/index.md
@@ -1,1 +1,0 @@
-{% include "dashboards/example/index.html" %}
