--- pages/dashboards/index.md
@@ -1,18 +1,0 @@
-# OSL Data Science Dashboard Gallery
-
-<div class="container mt-4">
-    <div class="row">
-
-        <div class="col-md-12 col-lg-6 col-xl-4">
-            <div class="card mb-4 p-0">
-                <img src="/images/dashboards/example/header.jpg" class="card-img-top my-0" alt="OSL Dashboard Example">
-                <div class="card-body">
-                    <h5 class="card-title">OSL Dashboard Example</h5>
-                    <p class="card-text">Example of datashboard to be used as reference</p>
-                    <a href="/dashboards/example/" class="btn btn-primary">View Dashboard</a>
-                </div>
-            </div>
-        </div>
-
-    </div>
-</div>
