+++ pages/projects/README.md
@@ -1,0 +1,3 @@
+The content of this folder is created automatically.
+
+Don't change it manually.
