+++ pages/projects/example/index.md
@@ -1,0 +1,21 @@
+---
+title: OSL Dashboard Example
+slug: example
+description: Example of dashboard to be used as reference
+
+authors:
+
+  - Ivan Ogasawara
+
+date: 2024-12-27
+template: dashboards.html
+---
+
+# OSL Dashboard Example (2024-12-27)
+
+#### Authors: Ivan Ogasawara
+
+Example of dashboard to be used as reference
+
+
+{% include "projects/example/index.html" %}
