+++ pages/projects/index.md
@@ -1,0 +1,27 @@
+# OSL Data Science Projects Gallery
+
+**Welcome to our Project Gallery!**
+
+Here, you can explore a variety of data science
+projects affiliated to the OSL Data Science Internship Program.
+Each project showcased here utilizes open data and is freely available as
+open-source, embodying our commitment to transparency and collaboration in
+science.
+
+<div class="container mt-4">
+    <div class="row">
+
+        <div class="col-md-12 col-lg-6 col-xl-4">
+            <div class="card mb-4 p-0">
+                <img src="/images/projects/example/header.jpg" class="card-img-top my-0" alt="OSL Dashboard Example">
+                <div class="card-body">
+                    <h5 class="card-title">OSL Dashboard Example</h5>
+                    <p class="card-text">Example of dashboard to be used as reference</p>
+                    <a href="/projects/example/" class="btn btn-primary">View Dashboard</a>
+                    <a href="https://github.com/osl-data-science/osl-data-science.github.io/tree/main/src/osl_data_science/projects/example/" class="btn btn-dark">Source Code</a>
+                </div>
+            </div>
+        </div>
+
+    </div>
+</div>
