--- src/osl_data_science/dashboards/__init__.py
@@ -1,1 +1,0 @@
-"""Dashboard package."""
