--- src/osl_data_science/dashboards/example/metadata.yaml
@@ -1,6 +1,0 @@
-title: "OSL Dashboard Example"
-slug: "example"
-# this folder is located at /theme folder
-image: "/images/dashboards/example/header.jpg"
-description: |
-  Example of datashboard to be used as reference
