+++ src/osl_data_science/projects/__init__.py
@@ -1,0 +1,1 @@
+"""Dashboard package."""
