+++ theme/dashboards.html
@@ -6,6 +6,7 @@
   color: #000000 !important;
 }
 </style>
+<script src="https://cdn.plot.ly/plotly-2.35.2.min.js" charset="utf-8"></script>
 {% endblock header_extra%}
 
 {% block content %}
