--- theme/images/dashboards/README.md
@@ -1,1 +1,0 @@
-Create a folder for your dashboard here and add all the images you need there.
