+++ theme/images/projects/README.md
@@ -1,0 +1,1 @@
+Create a folder for your dashboard here and add all the images you need there.
